more on smp, still wont boot

+++ debug.c
@@ -1,9 +1,8 @@
 #include "stdint.h"
-#include <stddef.h>
 #include "ports.h"
 
-static size_t strlen(const char* str) {
-	size_t len = 0;
+static uint32_t strlen(const char* str) {
+	uint32_t len = 0;
 	while (str[len])
 		len++;
 	return len;
